Making changes to existing server
- Made changes to Player structure
- Created Score service to manage game/player scores
- Created MapUtil to sort map as per values to use it for getting top scores

+++ server-client/game-server/src/main/java/com/game/engine/dao/PlayerDao.java
@@ -11,7 +11,7 @@ import java.util.Map;
  * Created by Sushant on 06-11-2016.
  */
 public class PlayerDao {
-    private static Integer id = 1;
+    private static Long id = 1L;
     private static Map<String, Player> players = new HashMap<>();
 
     public PlayerDao() {
